fix: StrategyEngine::activate() catch load_from_file exceptions
- Wrapped StrategyGraph::load_from_file() in try/catch to prevent
  SIGABRT on missing/corrupt graph files
- Backend now survives bad graph_path instead of crashing
- Verified: 3 strategies, 175 NAV, 3 trades, sharpe=0.116

+++ cpp/quant/api/strategy_api.cc
@@ -665,7 +665,7 @@ ApiResponse StrategyApi::trigger_backtest(uint64_t id, const std::string& body) 
                 } else if (k == "symbol") {
                     params.symbol = p.parse_string();
                 } else if (k == "kline_type") {
-                    params.kline_type = static_cast<uint8_t>(p.parse_number());
+                    params.kline_type = static_cast<int>(p.parse_number());
                 } else {
                     p.skip_value();
                 }
